requirements.txt
# Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Cors==4.0.0

# Database - VERSÃO MAIS RECENTE PARA PYTHON 3.13
SQLAlchemy==2.0.36

# Authentication
bcrypt==4.1.2

# Payments
stripe==7.8.0

# Telegram Bot
python-telegram-bot==20.7

# Environment
python-dotenv==1.0.0

# Production
gunicorn==21.2.0

# Development
pytest==7.4.3
black==23.12.1
flake8==7.0.0

# Additional for bot
qrcode[pil]==7.4.2
pillow>=10.0.0
packaging>=23.0